requirements.txt
###########################################################
# ⚙️ Core Framework
###########################################################
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic>=2.10.4
pydantic-settings==2.5.2

###########################################################
# 🧱 Database
###########################################################
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.14.0
python-dateutil==2.9.0.post0

###########################################################
# ☁️ Cloud & Storage
###########################################################
boto3==1.34.160          # Работа с MinIO / S3
python-multipart==0.0.9  # Загрузка файлов через API
python-magic-bin==0.4.14     # FIX: Windows-compatible libmagic

###########################################################
# 📄 Document & OCR Processing
###########################################################
pdfminer.six==20250506            # Извлечение текста из PDF (текстовые PDF, без OCR)
ocrmypdf==16.4.3                  # Автоматический OCR PDF (умный wrapper вокруг Tesseract + Ghostscript)
pytesseract==0.3.13               # OCR Tesseract для изображений (основной движок)
Pillow==10.4.0                    # Работа с изображениями (PIL): загрузка, конвертация, обработка
pdf2image==1.17.0                 # Конвертация PDF → изображения (JPEG/PNG) с помощью Poppler
python-docx==1.1.0                # Парсинг DOCX файлов (извлечение текста)
tqdm==4.67.1                      # Красивые прогресс-бары для потоковой обработки файлов
opencv-python-headless==4.12.0.88 # OpenCV без GUI: улучшение качества изображений перед OCR (preprocessing)
numpy==2.2.6                      # Совместимая версия с OpenCV


###########################################################
# 🧠 LLM / HTTP Clients
###########################################################
requests==2.32.3
httpx==0.27.2

###########################################################
# 🧾 PDF Generation & Export
###########################################################
reportlab==4.2.0         # Генерация PDF
pypdf==4.3.1             # Работа с PDF страницами (split/merge)

###########################################################
# ⚙️ Utilities
###########################################################
python-dotenv==1.0.1     # Работа с .env файлами
typing-extensions>=4.12.2
